Move notes into sub-package
