+++ src/main/java/org/opentripplanner/model/plan/StreetLegBuilder.java
@@ -5,7 +5,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import org.locationtech.jts.geom.LineString;
-import org.opentripplanner.street.model.StreetNote;
+import org.opentripplanner.street.model.note.StreetNote;
 import org.opentripplanner.street.search.TraverseMode;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
