+++ src/main/java/org/opentripplanner/model/plan/WalkStep.java
@@ -8,8 +8,8 @@ import java.util.Set;
 import org.opentripplanner.framework.i18n.I18NString;
 import org.opentripplanner.framework.lang.DoubleUtils;
 import org.opentripplanner.model.VehicleRentalStationInfo;
-import org.opentripplanner.street.model.StreetNote;
 import org.opentripplanner.street.model.edge.Edge;
+import org.opentripplanner.street.model.note.StreetNote;
 import org.opentripplanner.transit.model.basic.WgsCoordinate;
 
 /**
