+++ src/main/java/org/opentripplanner/routing/services/notes/StreetNotesSource.java
@@ -1,8 +1,8 @@
 package org.opentripplanner.routing.services.notes;
 
 import java.util.Set;
-import org.opentripplanner.street.model.StreetNoteAndMatcher;
 import org.opentripplanner.street.model.edge.Edge;
+import org.opentripplanner.street.model.note.StreetNoteAndMatcher;
 
 /**
  * A source of notes for edges.
