+++ src/main/java/org/opentripplanner/street/model/note/StreetNote.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.street.model;
+package org.opentripplanner.street.model.note;
 
 import java.io.Serializable;
 import java.util.Date;
